dev: developing mexc
