+++ tests/mexc/mexc_test.go
@@ -5,11 +5,11 @@ import (
 	"os"
 	"path/filepath"
 	"testing"
-	"time"
 
 	"github.com/Lazy-Parser/Collector/api"
 	"github.com/Lazy-Parser/Collector/config"
 	"github.com/Lazy-Parser/Collector/exchange"
+	"github.com/Lazy-Parser/Collector/market"
 )
 
 func TestMexc(t *testing.T) {
@@ -21,24 +21,43 @@ func TestMexc(t *testing.T) {
 		t.Fatalf("failed to setup config service. \nError: %v", err)
 	}
 
+	t.Log("Initing...")
 	mexcApi := api.NewMexcApi(cfg)
 	mexcE, err := exchange.NewMexc(mexcApi)
 	if err != nil {
 		t.Error(err)
 	}
 
+	ctx := context.Background()
+
+	// blocking
+	ch := make(chan *market.MexcSpotTick, 1024)
+	go func() {
+		if err := mexcE.ListenSpot(ctx, ch); err != nil {
+			t.Error(err)
+		}
+	}()
+
 	t.Log("Exchange created, try to fetch buffer")
-	if err := mexcE.BufferLoop(context.Background()); err != nil {
+	if err := mexcE.BufferLoop(ctx); err != nil {
 		t.Error(err)
 	}
-	t.Log("Fetched")
+	t.Log("Buffer Loop started success")
 
+	// listen ticks
 	for {
-		i := 1
-		for symbol, meta := range *mexcE.GetBuffer() {
-			t.Logf("%d) Symbol: %s | Meta: %+v", i, symbol, meta)
-			i++
+		select {
+		case <-ctx.Done():
+			return
+
+		case tick := <-ch:
+			t.Logf(
+				"%s: ASK %s | BID %s | Withdraw %t | Deposit %t",
+				tick.Symbol, tick.AskPrice, tick.BidPrice, tick.Withdraw, tick.Deposit,
+			)
 		}
-		time.Sleep(time.Second * 10)
 	}
+
+	close(ch)
+	return
 }
